requirements.txt
PyYAML>=3
keyring>=8
